# add Image.SubImageInto method